refactor(graphql): split es-api request-functions to new files
